Added contact, updated html's, began CSS border work
Added the contact.html, updated some classes in all the other html's, began working on border stylization for headers of pages, starting with the one for the GLORYHAMMER title itself. I am a bit confused as to 1. why it is not showing, and 2. how to get the text, and subsequently the border style, to wrap behind the image title.