Stylized pages
Added correct colors to content, added borders to page headers, reformatted Tour page. Need help configuring border size to be similar to that of my Figma file, as in the left side extends beyond the page, and has space on the right side, while being small enough on height to let the top and bottom of the contained text bleed through.