Cross-referenced separated edits
Cross-referenced edits made between GitHub version and desktop version of code, edited tour paragraphs to line up better. Personal list of changes that I know still need to be made: Still need to move Bio paragraphs and make them fill the page a bit better, extend the childpage header borders to almost the right end of the page while moving the text to justify near the left side of the page, move the back arrow to the top-left corner of the childpages (attempted to move it but something was finicky), apply border style to GLORYHAMMER title page header, center sub-head of title page and move it to the bottom.