Attempted border fix + additions
Attempted to fix the border of the main page title, managed to get the top and bottom to stick out but can't seem to get it to extend horizontally without changing the type size too. Flipped the footer of the main page to the correct hierarchy. Added fav icon code into every html file. Added media queries template to the end of the main CSS page under styles.